requirements.txt
atproto
peewee
Flask
python-dotenv
gunicorn
PyMySQL
tqdm
websocket-client
zstandard